+++ src/types/index.ts
@@ -1,14 +1,20 @@
-export type { ResolvedStoreBindings, ShallowStoreBindings } from "./store-bindings.types";
+export type { ShallowStoreBindings } from "../core/createShallowStore";
+export type {
+  GlobalStoreBindings,
+  StoreToolkit,
+  StoreToolkitOptions,
+} from "../core/createStoreToolkit";
+export type { ResolvedStoreBindings } from "../hooks/createResolvedStoreHooks";
+export type {
+  StoreProviderConfig,
+  StoreProviderProps,
+  StoreProviderResult,
+} from "../providers/createStoreProvider";
 export type { StorePlainHook, StoreValueHook } from "./store-hooks.types";
 export type {
   MutatorsStateCreator,
   SimpleStateCreator,
   StoreApiWithMutators,
+  StoreRecipe,
   StoreMutatorTuple,
 } from "./store.types";
-export type {
-  StoreProviderConfig,
-  StoreProviderProps,
-  StoreProviderResult,
-} from "./store-provider.types";
-export type { StoreToolkit } from "./store-toolkit.types";